+++ duckscript_sdk/src/sdk/internal/test/test_file/mod.rs
@@ -30,10 +30,6 @@ impl Command for CommandImpl {
         pckg::concat(&self.package, "TestFile")
     }
 
-    fn aliases(&self) -> Vec<String> {
-        vec!["test_file".to_string()]
-    }
-
     fn help(&self) -> String {
         "".to_string()
     }